requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
numpy==1.17.4
oauthlib==3.1.0
Pillow==6.2.1
requests==2.22.0
requests-oauthlib==1.3.0
schedule==0.6.0
urllib3==1.25.7
